Shiny Gadget for 'jsonedit'

Provides a Shiny gadget interface for jsonedit to interactively edit and return the changes for use in R.

jsonedit_gadget(..., height = NULL, width = NULL)

Arguments

...

arguments for jsonedit

height, width

any valid CSS size unit for the height and width of the gadget

Examples

if (FALSE) {
library(listviewer)

jsonedit_gadget(
  structure(
    as.list(1:4),
    names=letters[1:4]
  )
)
}