listviewer.

htmlwidget for interactive views of R lists

Details

R lists, especially nested lists, can be very difficult to visualize or represent. str just isn't enough, so this suite of htmlwidgets is designed to help see, understand, and maybe even modify your R lists.